backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "google-generativeai>=0.8.4",
    "openai>=1.72.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.0",
]
